Confusing test and validation concepts
According to GCP documentation (https://cloud.google.com/automl-tables/docs/prepare#ml-use) the split between train, validation and test is not properly explained in this example. Train and validation are the datasets that AutoML uses for training trial models and when a model is chosen then test dataset is shown to the model.